Split classes to own files
optimize string concatenate
