+++ src/ZPLUtility.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2005
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31229.75
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZPLUtility", "ZPLUtility\ZPLUtility.csproj", "{A836CA9D-CF58-4289-85F7-781EFF51E091}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZPLUtility", "ZPLUtility\ZPLUtility.csproj", "{A836CA9D-CF58-4289-85F7-781EFF51E091}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZPLUtilityUnitTest", "ZPLUtilityUnitTest\ZPLUtilityUnitTest.csproj", "{CE141155-ACCE-4F1E-BD86-49CC65E11302}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZPLUtility.UnitTest", "ZPLUtility.UnitTest\ZPLUtility.UnitTest.csproj", "{B385D358-B68A-4089-8FD9-6084D766FF8C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{A836CA9D-CF58-4289-85F7-781EFF51E091}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A836CA9D-CF58-4289-85F7-781EFF51E091}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A836CA9D-CF58-4289-85F7-781EFF51E091}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CE141155-ACCE-4F1E-BD86-49CC65E11302}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CE141155-ACCE-4F1E-BD86-49CC65E11302}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CE141155-ACCE-4F1E-BD86-49CC65E11302}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CE141155-ACCE-4F1E-BD86-49CC65E11302}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B385D358-B68A-4089-8FD9-6084D766FF8C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B385D358-B68A-4089-8FD9-6084D766FF8C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B385D358-B68A-4089-8FD9-6084D766FF8C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B385D358-B68A-4089-8FD9-6084D766FF8C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
