+++ src/ZPLUtility/GraphicSymbolCharacter.cs
@@ -1,0 +1,11 @@
+﻿namespace BinaryKits.Utility.ZPLUtility
+{
+    public enum GraphicSymbolCharacter
+    {
+        RegisteredTradeMark,
+        Copyright,
+        TradeMark,
+        UnderwritersLaboratoriesApproval,
+        CanadianStandardsAssociationApproval
+    }
+}
