+++ src/ZPLUtility/NewLineConversionMethod.cs
@@ -1,0 +1,9 @@
+﻿namespace BinaryKits.Utility.ZPLUtility
+{
+    public enum NewLineConversionMethod
+    {
+        ToSpace,
+        ToEmpty,
+        ToZPLNewLine,
+    }
+}
