+++ src/ZPLUtility/ZPLPositionedElementBase.cs
@@ -1,0 +1,12 @@
+﻿namespace BinaryKits.Utility.ZPLUtility
+{
+    public abstract class ZPLPositionedElementBase : ZPLElementBase
+    {
+        public ZPLOrigin Origin { get; protected set; }
+
+        public ZPLPositionedElementBase(int positionX, int positionY) : base()
+        {
+            Origin = new ZPLOrigin(positionX, positionY);
+        }
+    }
+}
